requirements.txt
tensorflow
matplotlib
seaborn
numpy
pygubu
tkinter
scipy
librosa
python-docx
sounddevice
